+++ Ordina.StichtingNuTwente/Ordina.StichtingNuTwente.WebApp/Controllers/GastgezinController.cs
@@ -17,16 +17,11 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
         private readonly IPlaatsingenService _plaatsingenService;
         private readonly IUserService _userService;
 
-        public GastgezinController(IGastgezinService gastgezinService,IPlaatsingenService plaatsingenService, IUserService userService, IMailService mailService)
+        public GastgezinController(IGastgezinService gastgezinService, IPlaatsingenService plaatsingenService, IUserService userService, IMailService mailService)
         {
             _gastgezinService = gastgezinService;
             _plaatsingenService = plaatsingenService;
             _userService = userService;
-        }
-
-        public IActionResult Overview()
-        {
-            return View();
         }
 
         [Route("gastgezin")]
@@ -96,7 +91,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
             viewModel.CanDelete = false;
             if (User.HasClaims("groups", "group-coordinator", "group-superadmin"))
                 viewModel.CanDelete = true;
-            
+
             return View(viewModel);
         }
 
@@ -114,7 +109,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
         }
 
         [HttpPost]
-        public IActionResult UpdateOpties(int GastGezinId, bool NoodOpvang, DateTime OnHoldTill ,bool OnHold, bool HasVOG, int MaxYoungerThanThree, int MaxOlderThanTwo)
+        public IActionResult UpdateOpties(int GastGezinId, bool NoodOpvang, DateTime OnHoldTill, bool OnHold, bool HasVOG, int MaxYoungerThanThree, int MaxOlderThanTwo)
         {
             var gastgezin = _gastgezinService.GetGastgezin(GastGezinId);
             if (gastgezin != null)
@@ -136,7 +131,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
                 gastgezin.HasVOG = HasVOG;
                 gastgezin.MaxOlderThanTwo = MaxOlderThanTwo;
                 gastgezin.MaxYoungerThanThree = MaxYoungerThanThree;
-                if(gastgezin.PlaatsingsInfo != null)
+                if (gastgezin.PlaatsingsInfo != null)
                 {
                     gastgezin.PlaatsingsInfo.VolwassenenGrotereKinderen = MaxOlderThanTwo.ToString();
                     gastgezin.PlaatsingsInfo.KleineKinderen = MaxYoungerThanThree.ToString();
@@ -154,145 +149,10 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
         {
             _userService.checkIfUserExists(User);
 
-            var model = new BeschikbareGastgezinnenModel();
+            var model = _gastgezinService.BeschikBeschikbareGastgezinnen(sortBy, sortOrder, filters, statusFilter, GetUser());
 
-            var gastgezinQuery = _gastgezinService.GetAllGastgezinnen().Where(g => g.IntakeFormulier != null);
-
-            if (!string.IsNullOrEmpty(statusFilter))
-            {
-                switch (statusFilter)
-                {
-                    case "Beschikbaar":
-                        gastgezinQuery = gastgezinQuery.Where(g => !g.NoodOpvang && g.Status == GastgezinStatus.Bezocht);
-                        break;
-                    case "Gereserveerd":
-                        gastgezinQuery = gastgezinQuery.Where(g => g.Status == GastgezinStatus.Gereserveerd);
-                        break;
-                    case "Geplaatst":
-                        gastgezinQuery = gastgezinQuery.Where(g => g.Status == GastgezinStatus.Geplaatst);
-                        break;
-                    case "Nood":
-                        gastgezinQuery = gastgezinQuery.Where(g => g.NoodOpvang);
-                        break;
-                    case "On Hold":
-                        gastgezinQuery = gastgezinQuery.Where(g => g.OnHold);
-                        break;
-                }
-            }
-
-            if (filters != null && filters.Length > 0)
-            {
-                var originalQuery = gastgezinQuery;
-                foreach (var filterParameter in filters)
-                {
-                    var split = filterParameter.Split('=');
-                    if (split.Length > 1)
-                    {
-                        var filterKey = split[0];
-                        var filterValue = split[1].ToLower();
-                        var results = 0;
-                        switch (filterKey)
-                        {
-                            case "Notitie":
-                                gastgezinQuery = gastgezinQuery.Where(g => g.Note != null && g.Note.ToLower().Contains(filterValue));
-                                results = originalQuery.Count(g => g.Note != null && g.Note.ToLower().Contains(filterValue));
-                                break;
-                            case "Opmerkingen":
-                                gastgezinQuery = gastgezinQuery.Where(g => g.VrijwilligerOpmerkingen != null && g.VrijwilligerOpmerkingen.ToLower().Contains(filterValue));
-                                results = originalQuery.Count(g => g.VrijwilligerOpmerkingen != null && g.VrijwilligerOpmerkingen.ToLower().Contains(filterValue));
-                                break;
-                            case "Buddy":
-                                gastgezinQuery = gastgezinQuery.Where(g => g.Buddy != null && (g.Buddy.FirstName.Contains(filterValue,StringComparison.CurrentCultureIgnoreCase)));
-                                results = originalQuery.Count(g => g.Buddy != null && (g.Buddy.FirstName.Contains(filterValue, StringComparison.CurrentCultureIgnoreCase)));
-                                break;
-                            default:
-                                gastgezinQuery = gastgezinQuery.Where(g => g.PlaatsingsInfo?.GetValueByFieldString(filterKey)?.ToLower().Contains(filterValue) == true);
-                                results = originalQuery.Count(g => g.PlaatsingsInfo?.GetValueByFieldString(filterKey)?.ToLower().Contains(filterValue) == true);
-                                break;
-                        }
-                        model.SearchQueries.Add(new SearchQueryViewModel() { OriginalQuery = filterParameter, Field = filterKey, SearchQuery = filterValue, Results = results });
-                    }
-                }
-            }
-
-            var gastGezinnen = gastgezinQuery;
-            var user = GetUser();
-
-            foreach (var gastGezin in gastGezinnen)
-            {
-                var plaatsingTag = _plaatsingenService.GetPlaatsingTag(PlacementType.Plaatsing, gastGezin);
-                var reserveTag = _plaatsingenService.GetPlaatsingTag(PlacementType.Reservering, gastGezin);
-                var gastgezinViewModel = GastgezinMapping.FromDatabaseToWebModel(gastGezin, user, plaatsingTag, reserveTag);
-                model.MijnGastgezinnen.Add(gastgezinViewModel);
-            }
-            if (sortOrder == "Ascending")
-            {
-                switch (sortBy)
-                {
-                    case "Woonplaats":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Woonplaats";
-                        break;
-                    case "Naam":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.Naam).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Naam";
-                        break;
-                    case "Geplaatst":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.PlaatsingTag).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Geplaatst (laag-hoog)";
-                        break;
-                    case "Gereserveerd":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.ReserveTag.Contains("HOLD")).ThenBy(g => g.ReserveTag).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Gereserveerd (laag-hoog)";
-                        break;
-                    case "AanmeldingsId":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.AanmeldFormulierId == null).ThenBy(g => g.AanmeldFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "AanmeldingsId (laag-hoog)";
-                        break;
-                    case "IntakeId":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.IntakeFormulierId == null).ThenBy(g => g.IntakeFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "IntakeId (laag-hoog)";
-                        break;
-                }
-            }
-            else if (sortOrder == "Descending")
-            {
-                switch (sortBy)
-                {
-                    case "Geplaatst":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenByDescending(g => g.PlaatsingTag).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Geplaatst (hoog-laag)";
-                        break;
-                    case "Gereserveerd":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenByDescending(g => g.ReserveTag).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "Gereserveerd (hoog-laag)";
-                        break;
-                    case "AanmeldingsId":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.AanmeldFormulierId == null).ThenByDescending(g => g.AanmeldFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "AanmeldingsId (hoog-laag)";
-                        break;
-                    case "IntakeId":
-                        model.MijnGastgezinnen = model.MijnGastgezinnen.OrderBy(g => g.PlaatsingTag.Contains("HOLD")).ThenBy(g => g.IntakeFormulierId == null).ThenByDescending(g => g.IntakeFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        model.SortDropdownText = "IntakeId (hoog-laag)";
-                        break;
-                }
-            }
-            model.TotalPlaatsingTag = _plaatsingenService.GetPlaatsingenTag(gastGezinnen.ToList(), PlacementType.Plaatsing);
-            model.TotalResTag = _plaatsingenService.GetPlaatsingenTag(gastGezinnen.ToList(), PlacementType.Reservering);
-            model.TotalMaxAdults = gastGezinnen.Sum(g => g.MaxOlderThanTwo);
-            model.TotalMaxChildren = gastGezinnen.Sum(g => g.MaxYoungerThanThree);
-            FillBaseModel(model);
+            _gastgezinService.FillBaseModel(model, _userService.getUserFromClaimsPrincipal(User));
             return View(model);
-        }
-
-        public void FillBaseModel(BaseModel model)
-        {
-            var user = _userService.getUserFromClaimsPrincipal(User);
-
-            if (user == null || user.Roles == null) return;
-
-            model.IsSecretariaat = user.Roles.Contains("group-secretariaat");
-            model.IsVrijwilliger = user.Roles.Contains("group-vrijwilliger");
         }
 
         [Authorize(Policy = "RequireCoordinatorRole")]
@@ -399,7 +259,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
 
         [HttpPut]
         [Route("RejectBuddy/{gastgezinId}")]
-        public IActionResult RejectBeingBuddy(int gastgezinId, string comment)
+        public IActionResult RejectBeingBuddy(int gastgezinId, string comment) // NIET GEBRUIKT
         {
             try
             {
@@ -443,7 +303,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
             if (user == null)
                 return Redirect("Error");
             var mijnGastgezinnen = FillMijnGastgezinnenModel(filter, user, editAddress);
-            FillBaseModel(mijnGastgezinnen);
+            _gastgezinService.FillBaseModel(mijnGastgezinnen, _userService.getUserFromClaimsPrincipal(User));
             return View(mijnGastgezinnen);
         }
 
@@ -478,7 +338,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
             mijnGastgezinnen.EditAddress = (bool)editAddress;
             }
 
-            FillBaseModel(mijnGastgezinnen);
+            _gastgezinService.FillBaseModel(mijnGastgezinnen, _userService.getUserFromClaimsPrincipal(User));
             return mijnGastgezinnen;
         }
 
@@ -489,115 +349,9 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
         public IActionResult AlleGastgezinnen(string? sortBy = "Woonplaats", string? sortOrder = "Ascending", string statusFilter = "")
         {
             _userService.checkIfUserExists(User);
+            var alleGastgezinnen = _gastgezinService.AlleGastgezinnen(sortBy, sortOrder, statusFilter, GetUser(), _userService.GetAllDropdownUsers().OrderBy(u => u.FirstName).ToList());
 
-            var alleGastgezinnen = new AlleGastgezinnenModel();
-
-            var vrijwilligers = _userService.GetAllDropdownUsers().OrderBy(u => u.FirstName).ToList();
-            foreach (var vrijwilliger in vrijwilligers.OrderBy(e => e.FirstName).ThenBy(e => e.LastName))
-            {
-                alleGastgezinnen.Vrijwilligers.Add(new Vrijwilliger(vrijwilliger));
-            }
-            var user = GetUser();
-            IEnumerable<Gastgezin> gastGezinnen = _gastgezinService.GetAllGastgezinnen();
-
-            if (!string.IsNullOrEmpty(statusFilter))
-            {
-                switch (statusFilter)
-                {
-                    case "Beschikbaar":
-                        gastGezinnen = gastGezinnen.Where(g => !g.NoodOpvang && g.Status == GastgezinStatus.Bezocht);
-                        break;
-                    case "Gereserveerd":
-                        gastGezinnen = gastGezinnen.Where(g => g.Status == GastgezinStatus.Gereserveerd);
-                        break;
-                    case "Geplaatst":
-                        gastGezinnen = gastGezinnen.Where(g => g.Status == GastgezinStatus.Geplaatst);
-                        break;
-                    case "Nood":
-                        gastGezinnen = gastGezinnen.Where(g => g.NoodOpvang);
-                        break;
-                    case "On Hold":
-                        gastGezinnen = gastGezinnen.Where(g => g.OnHold);
-                        break;
-                    case "Geen Intaker":
-                        gastGezinnen = gastGezinnen.Where(g => g.Intaker == null);
-                        break;
-                    case "Geen Buddy":
-                        gastGezinnen = gastGezinnen.Where(g => g.Buddy == null);
-                        break;
-                }
-            }
-
-            foreach (var gastGezin in gastGezinnen)
-            {
-                if (gastGezin.Contact == null)
-                {
-                    continue;
-                }
-                var plaatsingTag = _plaatsingenService.GetPlaatsingTag(PlacementType.Plaatsing, gastGezin);
-                var reserveTag = _plaatsingenService.GetPlaatsingTag(PlacementType.Reservering, gastGezin);
-                var gastgezinViewModel = GastgezinMapping.FromDatabaseToWebModel(gastGezin, user, plaatsingTag, reserveTag);
-                alleGastgezinnen.Gastgezinnen.Add(gastgezinViewModel);
-            }
-
-            if (sortOrder == "Ascending")
-            {
-                switch (sortBy)
-                {
-                    case "Woonplaats":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Woonplaats";
-                        break;
-                    case "Naam":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.Naam).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Naam";
-                        break;
-                    case "Telefoonnummer":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.Telefoonnummer).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Telefoonnummer";
-                        break;
-                    case "Intaker":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.Intaker).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Intaker (laag-hoog)";
-                        break;
-                    case "Buddy":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.Buddy).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Buddy (laag-hoog)";
-                        break;
-                    case "AanmeldingsId":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.AanmeldFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "AanmeldingsId (laag-hoog)";
-                        break;
-                    case "IntakeId":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderBy(g => g.IntakeFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "IntakeId (laag-hoog)";
-                        break;
-                }
-            }
-            else if (sortOrder == "Descending")
-            {
-                switch (sortBy)
-                {
-                    case "Intaker":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderByDescending(g => g.Intaker).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Intaker (hoog-laag)";
-                        break;
-                    case "Buddy":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderByDescending(g => g.Buddy).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "Buddy (hoog-laag)";
-                        break;
-                    case "AanmeldingsId":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderByDescending(g => g.AanmeldFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "AanmeldingsId (hoog-laag)";
-                        break;
-                    case "IntakeId":
-                        alleGastgezinnen.Gastgezinnen = alleGastgezinnen.Gastgezinnen.OrderByDescending(g => g.IntakeFormulierId).ThenBy(g => g.Woonplaats).ToList();
-                        alleGastgezinnen.SortDropdownText = "IntakeId (hoog-laag)";
-                        break;
-                }
-            }
-
-            FillBaseModel(alleGastgezinnen);
+            _gastgezinService.FillBaseModel(alleGastgezinnen, _userService.getUserFromClaimsPrincipal(User));
             return View(alleGastgezinnen);
         }
 
@@ -605,57 +359,9 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
         [HttpPost]
         public IActionResult IntakerOrBuddyChange(List<IntakerOrBuddyChangeModel> intakerOrBuddyChangeModels)
         {
-            var vrijwilligers = _userService.GetAllDropdownUsers().ToList();
-            foreach (var intakerOrBuddyChange in intakerOrBuddyChangeModels)
-            {
-                if (int.TryParse(intakerOrBuddyChange.Id, out int gastgezinId))
-                {
-                    var gastgezin = _gastgezinService.GetGastgezin(gastgezinId);
-                    if (gastgezin == null)
-                        return BadRequest();
-
-                    if (intakerOrBuddyChange.BuddyId != null)
-                    {
-                        if (intakerOrBuddyChange.BuddyId == "-")
-                        {
-                            gastgezin.Buddy = null;
-                        }
-                        else
-                        {
-                            if (int.TryParse(intakerOrBuddyChange.BuddyId, out int buddyId))
-                            {
-                                var buddy = vrijwilligers.FirstOrDefault(v => v.Id == buddyId);
-                                if (buddy != null && buddy.Id != gastgezin.Buddy?.Id)
-                                {
-                                    gastgezin.BekekenDoorBuddy = false;
-                                    gastgezin.Buddy = buddy;
-                                }
-                            }
-                        }
-                    }
-
-                    if (intakerOrBuddyChange.IntakerId != null)
-                    {
-                        if (intakerOrBuddyChange.IntakerId == "-")
-                        {
-                            gastgezin.Intaker = null;
-                        }
-                        else
-                        {
-                            if (int.TryParse(intakerOrBuddyChange.IntakerId, out int intakerId))
-                            {
-                                var intaker = vrijwilligers.FirstOrDefault(v => v.Id == intakerId);
-                                if (intaker != null && intaker.Id != gastgezin.Intaker?.Id)
-                                {
-                                    gastgezin.BekekenDoorIntaker = false;
-                                    gastgezin.Intaker = intaker;
-                                }
-                            }
-                        }
-                    }
-                    _gastgezinService.UpdateGastgezin(gastgezin, gastgezinId);
-                }
-            }
+            bool succes = _gastgezinService.IntakerOrBuddyChange(intakerOrBuddyChangeModels, _userService.GetAllDropdownUsers().ToList());
+            if (succes)
+                return BadRequest();
             return Ok();
         }
 
@@ -773,7 +479,7 @@ namespace Ordina.StichtingNuTwente.WebApp.Controllers
 
         [HttpGet]
         [AllowAnonymous]
-        public IActionResult GetGastgezinInformation()
+        public IActionResult GetGastgezinInformation() // NIET GEBRUIKT
         {
             var amountGastgezinnen = _gastgezinService.GetAllGastgezinnen().Where(g => g.Status == GastgezinStatus.Geplaatst);
             var amountGeplaatsteVluchtelingen = _plaatsingenService.GetPlaatsingen().Where(p => p.Active && (p.PlacementType == PlacementType.GeplaatsteReservering || p.PlacementType == PlacementType.Plaatsing));
