+++ Ordina.StichtingNuTwente/Ordina.StichtingNuTwente.WebApp/Models/ErrorViewModel.cs
@@ -3,7 +3,6 @@ namespace Ordina.StichtingNuTwente.WebApp.Models
     public class ErrorViewModel
     {
         public string? RequestId { get; set; }
-
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
     }
 }
